base entity is created for common attributes like timestamps and id
