+++ src/main/resources/note/todo.txt
@@ -1,7 +1,1 @@
-
-
-
 rearrange counts and get requests with consideration isActive
-
-
-check insert operation inserting object fields like inserting user with domain, restrict that with json ignore and check json ignore fields because some of them is not have I think.
